bug: otp-618: int to short convertion bug in AbstractBamFilePairAnalysisStartJob

+++ src/main/groovy/de/dkfz/tbi/otp/job/jobs/bamFilePairAnalysis/AbstractBamFilePairAnalysisStartJob.groovy
@@ -51,7 +51,7 @@ abstract class AbstractBamFilePairAnalysisStartJob extends AbstractStartJobImpl 
     @Override
     void execute() {
         SessionUtils.withNewSession {
-            short minPriority = minimumProcessingPriorityForOccupyingASlot
+            int minPriority = minimumProcessingPriorityForOccupyingASlot
             if (minPriority == ProcessingPriority.SUPREMUM) {
                 return
             }
